add logout, minor bug fixes, remove JWT on logout or delete

+++ frontend/main/main.js
@@ -4,6 +4,7 @@ const settingsButton = document.getElementById("settingsButton");
 const revertButton = document.getElementById("revertButton");
 const deleteButton = document.getElementById("deleteAccountButton");
 const settings = document.getElementById("settings");
+const logoutButton = document.getElementById("logoutButton");
 
 settingsButton.addEventListener("click", () => {
   settings.style.right = "0"; // Slide the panel in
@@ -13,6 +14,11 @@ settingsButton.addEventListener("click", () => {
 revertButton.addEventListener("click", () => {
   settings.style.right = "-250px"; // Slide the panel out
   revertButton.style.display = "none";
+});
+
+logoutButton.addEventListener("click", () => {
+  localStorage.removeItem("JWT");
+  window.location.href = "/frontend/login.html";
 });
 
 deleteButton.addEventListener("click", async () => {
@@ -26,17 +32,21 @@ deleteButton.addEventListener("click", async () => {
 });
 
 async function deleteAccount(userID) {
-  try {
-    const response = await axios.delete(
-      `http://localhost:4000/api/users/id/${userID}`
-    );
+  if (confirm("Are you sure you want to delete your account?")) {
+    try {
+      const response = await axios.delete(
+        `http://localhost:4000/api/users/id/${userID}`
+      );
+      console.log("ACCOUNT DELETED SUCCESSFUL");
+      localStorage.removeItem("JWT");
 
-    console.log("ACCOUNT DELETED SUCCESSFUL");
-  } catch (error) {
-    console.error(
-      "Error:",
-      error.response ? error.response.data : error.message
-    );
-    alert("An error occurred while trying to delete your account.");
+      window.location.href = "/frontend/main/accountDeleted.html";
+    } catch (error) {
+      console.error(
+        "Error:",
+        error.response ? error.response.data : error.message
+      );
+      alert("An error occurred while trying to delete your account.");
+    }
   }
 }
